backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
firebase-admin>=6.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
lxml>=4.9.0
